requirements.txt
crewai>=0.152.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
chromadb>=0.5.23
requests>=2.31.0
flask>=3.0.0 